move LessonLayout files to new folder
