+++ components/LessonLayout/index.tsx
@@ -1,0 +1,1 @@
+export { default, getLayout } from './LessonLayout'
